requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0
alembic>=1.13.0

# Authentication & Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Market Data
yfinance>=0.2.18
pandas>=2.2.0
numpy>=1.26.0

# Quantitative Finance (Core)
scipy>=1.12.0
scikit-learn>=1.4.0
statsmodels>=0.14.0

# Data Analysis & Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# API & Web
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0
selenium==4.15.2
coverage==7.3.2